Use concern to share password reset logic
