+++ app/models/artist.rb
@@ -1,11 +1,10 @@
 class Artist < ActiveRecord::Base
   include Activatable
-
-  attr_accessor :reset_token
-
-  has_many :grant_submissions
+  include PasswordReset
 
   has_secure_password
+
+  has_many :grant_submissions
 
   validates :name, :presence => true, length: { minimum: 4 }
   validates :email, :presence => true
@@ -17,26 +16,4 @@ class Artist < ActiveRecord::Base
     country = ISO3166::Country[contact_country]
     country.translations[I18n.locale.to_s] || country.name
   end
-
-  # These really should be private but then the password resetter can't get
-  # at them
-
-  # Sets the password reset attributes.
-  def create_reset_digest
-    self.reset_token = ApplicationController.new_token
-    update_attribute(:reset_digest,  ApplicationController.digest(reset_token))
-    update_attribute(:reset_sent_at, Time.zone.now)
-  end
-
-  # Sends password reset email.
-  def send_password_reset_email
-    UserMailer.password_reset("artists", self).deliver
-    logger.info "email: artist password reset sent to #{self.email}"
-  end
-
-  # Returns true if a password reset has expired.
-  def password_reset_expired?
-    reset_sent_at < 2.hours.ago
-  end
-
 end
